Add files via upload
It’s a fun little bash script. You feed it an Apple playlist, and it exports artists and albums, creating an HTML webpage with Google search links prioritized to Bandcamp. This way, by clicking on the links in the resulting HTML, you can go directly to the artists’ Bandcamp sites and buy the albums from your playlists, thus removing yourself from the cloud further. You have to export the Apple playlist and use it as a command line parameter with this script. I did it because the Apple Music buying experience sucks.